feat: 채팅 입력시 스크롤 조절
- 채팅 입력시 스크롤 조절 기능 추가
- 상세 채팅창 헤더 고정
- 스크롤 보이지 않도록 수정
