+++ client/src/app/globals.css
@@ -2,6 +2,17 @@
 @tailwind components;
 @tailwind utilities;
 
+
+.scroll-div{
+    -ms-overflow-style: none; /* IE and Edge */
+    scrollbar-width: none; /* Firefox */
+}
+
+.scroll-div::-webkit-scrollbar{
+    display: none;
+}
+
 body{
     @apply text-black
 }
+
