+++ client/src/components/layout/ChatingHeader.tsx
@@ -1,6 +1,6 @@
 export default function ChatingHeader() {
   return (
-    <header className="w-full p-4 border-b-2 border-solid sticky top-0 bg-white z-10">
+    <header className="w-full p-2 border-b-2 border-solid sticky top-0 bg-white z-10">
       <h2 className="font-bold text-2xl">채팅</h2>
     </header>
   );
